+++ gnumed/gnumed/test-area/java/gnmed/java/org/gnumed/gmGIS/street.java
@@ -23,7 +23,7 @@ public class street {
  * Represents ...
  * </p>
  */
-    private Integer id; 
+    private Long id; 
 
 /**
  * <p>
@@ -76,7 +76,7 @@ public class street {
  *  @hibernate.id
  *      generator-class="hilo"
  */
-    public Integer getId() {        
+    public Long getId() {        
         return id;
     } // end getId        
 
@@ -85,7 +85,7 @@ public class street {
  * Represents ...
  * </p>
  */
-    public void setId(Integer _id) {        
+    public void setId(Long _id) {        
         id = _id;
     } // end setId
