+++ gnumed/gnumed/test-area/java/gnmed/java/org/gnumed/gmIdentity/identity.java
@@ -38,7 +38,7 @@ public class identity {
      * Represents ...
      * </p>
      */
-    private Integer id;
+    private Long id;
     
     /**
      * <p>
@@ -417,7 +417,7 @@ public class identity {
      *  @hibernate.id
      *      generator-class="hilo"
      */
-    public Integer getId() {
+    public Long getId() {
         return id;
     } // end getId
     
@@ -426,7 +426,7 @@ public class identity {
      * Represents ...
      * </p>
      */
-    public void setId(Integer _id) {
+    public void setId(Long _id) {
         id = _id;
     }
